ETL/pyproject.toml
[project]
name = "etl"
version = "0.2.0"
description = "ETL pipeline as a Microservice"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.0",
    "falkordb",
    "fastapi",
    "httpx",
    "pymongo",
    "pytest",
    "python-dotenv",
    "redis",
    "uvicorn",
]
